Introduce primitive data types
Improves type safety when using "primitive" types and enables target
type resolution for specific Parquet, etc. types when performing model
introspection.

The abstract data types include Pydantic integer bounds that surface
in generated JSON Schema.

+++ packages/overture-schema-validation/src/overture/schema/validation/types.py
@@ -4,6 +4,8 @@ from typing import Annotated, NewType
 
 from pydantic import Field
 
+from overture.schema.core.primitives import float64
+
 from .constraints import (
     CategoryPatternConstraint,
     ConfidenceScoreConstraint,
@@ -63,7 +65,7 @@ WikidataId = NewType(
 ConfidenceScore = NewType(
     "ConfidenceScore",
     Annotated[
-        float,
+        float64,
         ConfidenceScoreConstraint(),
         Field(
             description="Confidence value from the source dataset, particularly relevant for ML-derived data."
